@@ -13,12 +13,12 @@
     "test": "jest"
   },
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
   },
   "devDependencies": {
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
     "@typescript-eslint/eslint-plugin": "^5.10.2",
     "@typescript-eslint/parser": "^5.10.2",
     "@vitejs/plugin-react": "^1.3.2",
@@ -27,11 +27,11 @@
     "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-jsx-a11y": "^6.5.1",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react": "^7.34.2",
     "eslint-plugin-simple-import-sort": "^7.0.0",
     "jest": "^29.7.0",
     "pre-commit": "^1.2.2",
-    "prettier": "^2.5.1",
+    "prettier": "^3.3.1",
     "typescript": "^4.5.5",
     "vite": "^5.2.13"
   },
